requirements.txt
pandas~=2.2.2
jira
pytz~=2024.1
requests~=2.32.3
python-dotenv~=1.0.1
